CNG module configuration in plot_util
