+++ src/malet/plot.py
@@ -19,8 +19,9 @@ from rich.panel import Panel
 from rich.columns import Columns
 from rich.align import Align
 
-from .plot_utils.metric_drawer import *
-from .plot_utils.utils import *
+from .plot_utils.data_processor import avgbest_df, select_df
+from .plot_utils.plot_drawer import ax_draw_curve, ax_draw_best_stared_curve, ax_draw_bar, ax_draw_heatmap
+from .plot_utils.utils import merge_dict, default_style, ax_styler, save_figure
 
 FLAGS = flags.FLAGS
